test: add CI test for tail calls hooks for custom programs
Add a new CI test to check that custom programs loaded and run via the
datapath tail call hooks work as expected. In particular, we use the
byte counter example provided in a previous commit.

The test is the following: We deploy Cilium and some pods, we load the
byte counter program and attach it to a hook (ingress or egress) for a
given endpoint, we send some traffic (ping requests and replies), and
then check the values in the eBPF map where the count is stored.

A new manifest is added, to make sure that we have two simple
application endpoints, located on different nodes. This is to check that
the egress hook is working (there is no egress hook for socket-based
load-balancing yet).